+++ sms-backup/src/main/res/values/strings.xml
@@ -7,4 +7,8 @@
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="str_btn_backup">Backup</string>
+    <string name="str_btn_restore">Restore</string>
+    <string name="str_sms_backup_title">SMS Backup</string>
+    <string name="str_sms_backup_schedule_backups">Schedule Backups</string>
 </resources>
